feat(editor): replace one-click share with full share dialog
The editor toolbar's Share button used to POST to the share endpoint,
copy the URL to the clipboard, and that was it. Every click also
re-rolled the share ID, silently invalidating any previously-shared
links — so opening it twice broke whatever embed someone had already
pasted into their docs.

The dashboard already had a proper share dialog (link + embed snippet
+ revoke). This pulls that body out into a shared ShareDialog
component and points both the dashboard wrapper and the editor
toolbar at it.

Editor flow now:
1. Click Share → dialog opens, lazily fetches the workflow to learn
   the existing shareId
2. If already public: shows the /w/<id> link, the iframe embed
   snippet, both with copy buttons, plus a Make Private to revoke
3. If private: shows a single Make Public action that generates a
   share ID and updates the dialog in place
